test(http): Extract session tests from `ApplicationTest` to `ApplicationSessionTest` class.
